+++ core/worker/script/index.js
@@ -1,10 +1,10 @@
 "use strict";
 
-const FAILURE_STATE = 'failure';
-const NORMAL_STATE = 'normal';
-
 var Worker = require('../index').define('script');
 var Script = require(APP_ROOT + '/lib/script');
+
+var FAILURE_STATE = 'failure';
+var SUCCESS_STATE = 'normal';
 
 module.exports = Worker;
 
@@ -22,26 +22,15 @@ Worker.prototype.initialize = function(){
 }
 
 Worker.prototype.getData = function(next) {
-  this.checkScript(this.script, error => {
+  var self = this;
+  this.checkScript(this.script,function(error){
     // if(error) return done(error);
-    this.script.run().end(result=>{
-    this.debug.log('result is %j',result);
+    self.script.run(function(result){
+      self.debug.log('result is %j',result);
       var lastline = result.lastline;
-      var data = { 'data':result };
-
-      var isEvent = isEventObject(lastline);
-      if( isEvent ) {
-        data.state = lastline.event;
-        data.event = lastline;
-      } else {
-        data.state = lastline;
-      }
-
+      var data = { 'data': result };
+      data.state = lastline;
       return next(null,data);
     });
   });
-}
-
-function isEventObject(input){
-  return (input instanceof Object) && input.event ;
 }